+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [Unreleased]
 - Update PHP CS Fixer to v2.16
 - Add DataProviderStaticFixer
+- Add NoSuperfluousConcatenationFixer
 - Add PhpdocTypesTrimFixer
+- Deprecate NoUnneededConcatenationFixer
 - Deprecate NullableParamStyleFixer
 - Deprecate SingleLineThrowFixer
 - Allow symfony/finder 5.0